[RB-05] - alteração do teste de Sala

+++ Backend/src/main/java/com/src/servico/SalaServico.java
@@ -1,8 +1,15 @@
 package com.src.servico;
 
+import com.src.dominio.Equipamento;
 import com.src.dominio.Sala;
 import com.src.dominio.SalaEquipamento;
+import com.src.repositorio.EquipamentoRepositorio;
+import com.src.repositorio.ReservaRepositorio;
+import com.src.servico.dto.ReservaDTO;
+import com.src.servico.dto.SalaEquipamentoDTO;
 import com.src.servico.excecao.RegraNegocioException;
+import com.src.servico.mapper.EquipamentoMapper;
+import com.src.servico.mapper.SalaEquipamentoMapper;
 import com.src.servico.mapper.SalaMapper;
 import com.src.repositorio.SalaEquipamentoRepositorio;
 import com.src.repositorio.SalaRepositorio;
@@ -19,14 +26,17 @@ import java.util.List;
 @Transactional
 public class SalaServico {
 
-    private final SalaEquipamentoRepositorio salaEquipamentoRepositorio;
     private final SalaRepositorio salaRepositorio;
+    private final SalaEquipamentoRepositorio salaEquipamentoRepositorio;
     private final SalaMapper salaMapper;
+    private final EquipamentoRepositorio equipamentoRepositorio;
+    private final ReservaRepositorio reservaRepositorio;
+    private final ReservaServico reservaServico;
+    private final EquipamentoServico equipamentoServico;
 
     public List<SalaDTO> listar(){
         List<Sala> salas = salaRepositorio.findAll();
-        List<SalaDTO> salasDTO = salaMapper.toDto(salas);
-        return salasDTO;
+        return salaMapper.toDto(salas);
     }
 
     public SalaDTO buscar(Integer id){
@@ -36,25 +46,44 @@ public class SalaServico {
         return salaDTO;
     }
 
-    public SalaDTO inserir(SalaDTO salaDTO){
-        if (salaDTO.getId() != null){
-            Sala sala = salaRepositorio.findById(salaDTO.getId())
-                    .orElseThrow(() -> new RegraNegocioException("Sala não encontrada"));
+    public SalaDTO inserir(SalaDTO salaDTO) throws RegraNegocioException{
+        Sala novaSala = salaMapper.toEntity(salaDTO);
+        List<SalaEquipamento> equipamentos = novaSala.getEquipamentos();
+        novaSala.setEquipamentos(new ArrayList<>());
+
+        for (SalaDTO salaInst: listar()){
+            if (salaInst.getDescricao().equals(novaSala.getDescricao()) || salaInst.getEquipamentos().equals(salaDTO.getEquipamentos())){
+                throw new RegraNegocioException("Salas iguais");
+            }
         }
-        Sala sala = salaMapper.toEntity(salaDTO);
-        List<SalaEquipamento> equipamentos = sala.getEquipamentos();
-        sala.setEquipamentos(new ArrayList<>());
-        salaRepositorio.save(sala);
-        equipamentos.forEach(equipamento -> {
-            equipamento.setSala(sala);
-            equipamento.getId().setIdSala(sala.getId());
+        salaRepositorio.save(novaSala);
+        equipamentos.forEach(equipamento ->{
+            equipamento.setSala(novaSala);
+            equipamento.getId().setIdSala(novaSala.getId());
         });
         salaEquipamentoRepositorio.saveAll(equipamentos);
-        return salaMapper.toDto(sala);
+        SalaDTO salaDTO1 = salaMapper.toDto((novaSala));
+        return salaDTO1;
     }
 
     public void deletar(Integer id){
-        salaRepositorio.findById(id).orElseThrow(() -> new RegraNegocioException(("Sala não encontrada")));
-        salaRepositorio.deleteById(id);
+        Sala sala = salaRepositorio.findById(id).orElseThrow(() -> new RegraNegocioException("Sala com o id " + id + " não existe"));
+        if (sala.getDisponivel() == 0) {
+            throw new RegraNegocioException("Esta sala não pode ser deletada devido a estar reservada.");
+        }
+        else {
+            salaEquipamentoRepositorio.deleteAllBySalaId(id);
+            salaRepositorio.deleteById(id);
+        }
     }
+
+    public boolean isReservada(Integer id){
+        for (ReservaDTO reserva: reservaServico.listar()) {
+            if(reserva.getIdSala() == id){
+                return true;
+            }
+        }
+        return false;
+    }
+
 }
